requirements.txt
asgiref==3.8.1
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dj-database-url==2.1.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
sqlparse==0.4.4
typing_extensions==4.10.0
whitenoise==6.6.0
PyJWT==2.8.0
pillow==10.2.0
